requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
mysql-connector-python==8.2.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy>=2.1.0
Pillow==10.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
